split table into separate crate
