+++ common/src/lib.rs
@@ -1,0 +1,3 @@
+mod error;
+
+pub use error::{Error, Result};
