+++ src/lib.rs
@@ -1,17 +1,12 @@
-mod checksum;
 mod db;
 mod entry;
-mod error;
 pub mod format;
 mod levels;
 mod memtable;
 mod ops;
-mod opt;
-// mod table;
 mod util;
 mod value;
 mod wal;
 
 pub use db::{Agate, AgateOptions};
-pub use error::{Error, Result};
 pub use skiplist::Skiplist;
