+++ table/src/opt.rs
